backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.66
botocore==1.42.66
click==8.3.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==25.1.0
h11==0.16.0
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
